+++ Assets/Scripts/Network/Packet/PacketType/PacketByte.cs
@@ -1,6 +1,6 @@
 namespace Network.Packet.PacketType {
 	public sealed class PacketByte: PacketType<byte> {
-		public override int? GetFixedSize() => sizeof(byte);
+		public override ushort? GetFixedSize() => sizeof(byte);
 		
 		public PacketByte() {}
 		public PacketByte(byte[] bytes): base(bytes) { }
